Moved vectors, matrices, and some math to geometry.h
