+++ CppSundial/src/base/baseImpl.cpp
@@ -4,6 +4,7 @@
 #include "hashtable.h"
 #include "allocators.h"
 #include "utils.h"
+#include "geometry.h"
