+++ CppSundial/src/colors.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "base/utils.h"
+#include "base/geometry.h"
 
 V4f col_darkBlue = V4f(27, 28, 37, 255) / 255;
 V4f col_darkGray = V4f(38, 40, 47, 255) / 255;
